[feat] better run settings

+++ set_tx2_reader_env.sh
@@ -10,15 +10,18 @@ fi
 START_DIR="${PWD}"
 JETSON_SETTINGS="${HOME}/jetson.conf"
 
-echo "Setting TX2 Runtime Environment"
+echo "Setting Runtime Environment"
 # set network params
 sysctl -w net.core.rmem_max=24266666
 sysctl -w net.core.wmem_max=24266666
 
-# turn max cpu performance on
-echo "See cpu info with /home/nvidia/tegrastats"
-su -c "echo 1 > /sys/devices/system/cpu/cpu1/online"
-su -c "echo 1 > /sys/devices/system/cpu/cpu2/online"
-rm -rf "${JETSON_SETTINGS}" # remove settings so dont ask if okay to overwrite
-/home/nvidia/jetson_clocks.sh --store "${JETSON_SETTINGS}" # save settings to restore from
-/home/nvidia/jetson_clocks.sh
+# make sure this is a tx2
+if [ -f /home/nvidia/jetson_clocks.sh ]; then
+    # turn max cpu performance on
+    echo "See cpu info with /home/nvidia/tegrastats"
+    su -c "echo 1 > /sys/devices/system/cpu/cpu1/online"
+    su -c "echo 1 > /sys/devices/system/cpu/cpu2/online"
+    rm -rf "${JETSON_SETTINGS}" # remove settings so dont ask if okay to overwrite
+    /home/nvidia/jetson_clocks.sh --store "${JETSON_SETTINGS}" # save settings to restore from
+    /home/nvidia/jetson_clocks.sh
+fi
